requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
motor==3.3.2
beanie==1.22.1
pymongo==4.6.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.8
pypdf==3.17.4
python-docx==1.1.0
openai==1.3.0
sentence-transformers==2.2.2
sib-api-v3-sdk==7.6.0 
python-dotenv==1.0.0
pydantic-settings==2.1.0
tenacity==8.2.3
psutil==5.9.6
python-json-logger==2.0.7
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.12.1
isort==5.13.2